+++ src/main/java/com/example/demo/Repository/UsuarioRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.math.BigInteger;
 
-public interface UsuarioRepository extends JpaRepository<Usuario, BigInteger> {
+public interface UsuarioRepository extends JpaRepository<Usuario, Long> {
 }
