+++ src/main/java/com/example/demo/Service/CredencialService.java
@@ -6,9 +6,9 @@ import java.math.BigInteger;
 import java.util.List;
 
 public interface CredencialService {
-    public Credencial getById(BigInteger id);
+    public Credencial getById(Long id);
     public List<Credencial> getAll();
     public Credencial save(Credencial credencial);
-    public Credencial update(BigInteger id, Credencial credencial);
-    public void delete(BigInteger id);
+    public Credencial update(Long id, Credencial credencial);
+    public void delete(Long id);
 }
